Improve decode auth token.
