+++ packages/api/src/main.ts
@@ -4,4 +4,4 @@ export * from './makeServices'
 export * from './makeMergedSchema/makeMergedSchema'
 export * from './functions/graphQLServer'
 export * from './globalContext'
-export * from './auth/getUserFromContext'
+export * from './auth/decodeAuthToken'
